refactor: Remove the bench module from common
