+++ Cargo.toml
@@ -16,20 +16,18 @@ path = "src/lib.rs"
 
 [dependencies]
 bitvec = "1.0.1"
-mt19937 = { version = "2.0.1", optional = true }
-rand = { version = "0.8.5", optional = true }
 thiserror = "1.0.63"
 
 [dev-dependencies]
 criterion = "0.5.1"
 cxx = "1.0.128"
+mt19937 = "2.0.1"
+rand = "0.8.5"
 
 [build-dependencies]
 cxx-build = "1.0.128"
 
 [features]
-bench = ["dep:mt19937", "dep:rand"]
-common = ["bench"]
 correctness = []
 
 [[bench]]
