+++ scripts/run_benchmarks.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-cargo bench --features bench
+cargo bench
