+++ src/lib.rs
@@ -44,9 +44,6 @@ mod shisanyao;
 mod shoupai;
 mod standard;
 
-#[doc(hidden)]
-pub mod common;
-
 pub use bingpai::Bingpai;
 pub use calculate::{calculate_replacement_number, calculate_replacement_number_3_player};
 pub use fulu_mianzi::{ClaimedTilePosition, FuluMianzi, InvalidFuluMianziError};
